Create Visual Studio Project
1) Why is the change being made?
---------------------------------
This repository will hold a C++ Visual Studio Project; this is the
barebones creation of that project to be built upon in later commits.

2) What has changed?
---------------------------------
Command line Visual Studio C++ project created which opens up a windows
terminal to print Hello World. Contains Precompiled Headers and a VS
.gitignore file.

3) Other concerns and/or dependent changes?
---------------------------------
N/A

4) Testing
---------------------------------
Hello World successfully runs.